user-workspace/banking-app/requirements.txt
# Core
flask==3.0.2
python-dotenv==1.0.0

# Security
flask-jwt-extended==4.5.3
flask-limiter==3.5.0
flask-talisman==0.8.1
bcrypt==4.0.1
pyjwt==2.8.0
cryptography==41.0.3
bleach==6.0.0

# Database
sqlalchemy==2.0.19
psycopg2-binary==2.9.7
alembic==1.12.0

# Machine Learning
scikit-learn==1.3.0
pandas==2.1.1
numpy==1.26.0

# Monitoring
sentry-sdk==1.34.0
prometheus-flask-exporter==0.22.4

# Testing
pytest==7.4.2
bandit==1.7.5
safety==2.3.5